requirements.txt
pandas
tqdm
numpy
scikit-learn
sentence-transformers
faiss-cpu
openai
pathway
python-dotenv
